requirements.txt
CoAPthon3
cbor2
noiseprotocol
requests
zstd

